Refactor: Remove compiled models for Family, FamilyMember, Member, Report, and Summary
- Deleted auto-generated compiled model classes for Family, FamilyMember, Member, Report, and Summary to streamline the codebase.
- Updated DonationConfiguration to use enum directly for default value.
- Added ServiceDefaults project with health check extensions.
- Introduced integration tests for donation flow, including creation and voiding of donations.
- Updated solution structure to include new projects and test configurations.
- Enhanced README with instructions for running the application and tests.
- Removed unnecessary compiled model generation steps from build scripts.

+++ ChurchApp.Application/ChurchApp.Application/Infrastructure/Configurations/Donations/DonationConfiguration.cs
@@ -11,7 +11,7 @@ public class DonationConfiguration : IEntityTypeConfiguration<Donation>
         builder.HasKey(x => x.Id);
         builder.Property(x => x.Type).HasConversion<int>();
         builder.Property(x => x.Method).HasConversion<int>();
-        builder.Property(x => x.Status).HasConversion<int>().HasDefaultValue((int)DonationStatus.Active);
+        builder.Property(x => x.Status).HasConversion<int>().HasDefaultValue(DonationStatus.Active);
         builder.Property(x => x.Amount).HasPrecision(18, 2);
         builder.Property(x => x.IdempotencyKey).HasMaxLength(100);
         builder.Property(x => x.ServiceName).HasMaxLength(200);
